move keywords
