+++ packages/crypto-validation/source/keywords.test.ts
@@ -177,72 +177,4 @@ describe<{
 		assert.false(object.amount instanceof BigNumber);
 		assert.equal(object.amount, "12");
 	});
-
-	it("keyword transactionGasLimit should be ok", (context) => {
-		const schema = {
-			$id: "test",
-			transactionGasLimit: {},
-		};
-		context.validator.addSchema(schema);
-
-		assert.undefined(context.validator.validate("test", cryptoJson.milestones[0].gas!.minimumGasLimit).error);
-		assert.undefined(context.validator.validate("test", cryptoJson.milestones[0].gas!.maximumGasLimit).error);
-
-		assert.defined(context.validator.validate("test", 1).error);
-		assert.defined(context.validator.validate("test", 0).error);
-		assert.defined(context.validator.validate("test", -1).error);
-		assert.defined(context.validator.validate("test", Number.MAX_SAFE_INTEGER).error);
-		assert.defined(context.validator.validate("test", "asdf").error);
-	});
-
-	it("keyword bytecode should be ok", (context) => {
-		const schema = {
-			$id: "test",
-			bytecode: {},
-		};
-		context.validator.addSchema(schema);
-
-		assert.undefined(context.validator.validate("test", "").error);
-		assert.undefined(context.validator.validate("test", "0x00").error);
-		assert.undefined(context.validator.validate("test", "0x").error);
-		assert.undefined(context.validator.validate("test", "00").error);
-
-		const maxBytecodeLength = cryptoJson.milestones[0].gas!.maximumGasLimit / 2;
-		const maxPayload = "0x" + "a".repeat(maxBytecodeLength);
-		assert.undefined(context.validator.validate("test", maxPayload).error);
-
-		assert.defined(context.validator.validate("test", maxPayload + "aa").error);
-
-		assert.defined(context.validator.validate("test", 1).error);
-		assert.defined(context.validator.validate("test", 0).error);
-		assert.defined(context.validator.validate("test", -1).error);
-		assert.defined(context.validator.validate("test", Number.MAX_SAFE_INTEGER).error);
-		assert.defined(context.validator.validate("test", "asdf").error);
-	});
-
-	it("keyword bytecode should remove 0x prefix", (context) => {
-		const schema = {
-			$id: "test",
-			type: "object",
-			properties: {
-				payload: { bytecode: {} },
-			},
-		};
-
-		context.validator.addSchema(schema);
-
-		const withPrefix = {
-			payload: "0xdead",
-		};
-
-		assert.undefined(context.validator.validate("test", withPrefix).error);
-		assert.equal(withPrefix.payload, "dead");
-
-		const withoutPrefix = {
-			payload: "dead",
-		};
-
-		assert.undefined(context.validator.validate("test", withoutPrefix).error);
-		assert.equal(withoutPrefix.payload, "dead");
-	});
 });
